+++ common/src/main/java/it/crystalnest/fancy_entity_renderer/api/entity/player/FancyPlayerRenderer.java
@@ -124,6 +124,9 @@ public class FancyPlayerRenderer extends PlayerRenderer {
       float offsetY = (state.pose == Pose.SLEEPING || state.pose == Pose.SWIMMING ? state.boundingBoxWidth : height) / scale + (float) state.nameTagAttachment.y;
       float offsetX = state.pose == Pose.SLEEPING ? -(float) state.nameTagAttachment.x : font.width(nameTag) / 2F;
       poseStack.scale(scale, -scale, scale);
+      if (state.pinName) {
+        poseStack.rotateAround(new Quaternionf().rotateY(state.modelRot.getY()), 0, 0, 0);
+      }
       poseStack.translate(-offsetX, -offsetY, 0);
       if (renderState.isUpsideDown) {
         poseStack.scale(1, -1, 1);
@@ -184,9 +187,15 @@ public class FancyPlayerRenderer extends PlayerRenderer {
         float step = renderState.speedValue * 0.33F;
         renderState.ageInTicks += step;
         renderState.walkAnimationPos += step;
+        if (!(state.pose == Pose.SPIN_ATTACK || state.pose == Pose.SLEEPING)) {
+          renderState.walkAnimationSpeed = state.walkSpeed;
+        } else {
+          renderState.walkAnimationSpeed = 0;
+        }
       } else {
         renderState.ageInTicks = 3000;
         renderState.walkAnimationPos = 0;
+        renderState.walkAnimationSpeed = 0;
       }
       renderState.nameTag = Component.literal(state.name);
       if (state.pose == Pose.SLEEPING) {